third_party/blink/web_tests/platform/mac/external/wpt/html/canvas/offscreen/text/2d.text.drawing.style.letterSpacing.measure.relative.worker-expected.txt

This is a testharness.js-based test.
FAIL Testing letter spacing with font-relative length assert_approx_equals: letter spacing doesn't work. expected 132.021484375 +/- 0.1 but got 66.0107421875
Harness: the test ran to completion.
